refactor(store): 重构boss和stock模块，合并功能到boss模块
- 将stock模块的功能迁移到boss模块
- 删除stock相关页面和组件
- 更新api调用路径和store引用
- 优化表格标题生成逻辑
- 调整首页添加实时业绩展示
