+++ components/summary/boss/Card.vue
@@ -22,17 +22,19 @@ const toggleChart = ref<'list' | 'chart'>('list')
       </template>
       <template #select>
         <template v-if="props.where">
-          <n-radio-group
-            v-model:value="radio" name="radiogroup" @update:value="emits('getlist')">
-            <n-radio
-              v-for="(item, index) in props.where?.type?.preset" :key="item" :value="Number(index)" :style="{
-                '--n-box-shadow-hover': 'inset 0 0 0 1px #0068ff',
-                '--n-box-shadow-active': 'inset 0 0 0 1px #0068ff',
-                '--n-dot-color-active': '#0068ff',
-                '--n-box-shadow-focus': 'inset 0 0 0 1px #0068ff, 0 0 0 2px rgba(24, 65, 160, 0.2)' }">
-              {{ item }}
-            </n-radio>
-          </n-radio-group>
+          <div class="pt-[12px]">
+            <n-radio-group
+              v-model:value="radio" name="radiogroup" @update:value="emits('getlist')">
+              <n-radio
+                v-for="(item, index) in props.where?.type?.preset" :key="item" :value="Number(index)" :style="{
+                  '--n-box-shadow-hover': 'inset 0 0 0 1px #0068ff',
+                  '--n-box-shadow-active': 'inset 0 0 0 1px #0068ff',
+                  '--n-dot-color-active': '#0068ff',
+                  '--n-box-shadow-focus': 'inset 0 0 0 1px #0068ff, 0 0 0 2px rgba(24, 65, 160, 0.2)' }">
+                {{ item }}
+              </n-radio>
+            </n-radio-group>
+          </div>
         </template>
       </template>
     </summary-table>
