+++ pages/index/index.vue
@@ -23,6 +23,28 @@ const handleSelectFn = async (id: Stores['id']) => {
 
 await myStoreTodayInventory({ store_id: myStore.value.id })
 await myStoreTodaySale({ store_id: myStore.value.id })
+
+// 性能统计初始化
+const { getPerformanceType, getPerformanceList } = useBoss()
+const params = ref({ duration: 1 } as BossWhere)
+const PerformanceLoading = ref<boolean>(false)
+const performanceTitle = ref<StockTitle[]>([])
+const performanceList = ref<BossSalesList[]>([])
+PerformanceLoading.value = true
+const getPerformanceListFn = async () => {
+  const res = await getPerformanceList({ ...params.value })
+  performanceTitle.value = res?.title || []
+  performanceList.value = res?.list || []
+  PerformanceLoading.value = false
+}
+
+const getPerformanceReq = async () => {
+  await getPerformanceType()
+  await getPerformanceListFn()
+}
+onMounted(async () => {
+  await getPerformanceReq()
+})
 </script>
 
 <template>
@@ -50,6 +72,12 @@ await myStoreTodaySale({ store_id: myStore.value.id })
         <template v-if="TodayInventory">
           <summary-card-inventory :today-inventory="TodayInventory" />
         </template>
+        <summary-boss-card
+          card-title="实时业绩"
+          :title="performanceTitle"
+          :list="performanceList"
+          :loading="PerformanceLoading"
+          @getlist="getPerformanceListFn" />
       </template>
       <template v-else>
         <common-emptys text="暂未分配门店" />
