+++ stores/home.ts
@@ -1,10 +1,8 @@
 export const homeDataStore = defineStore('homeDataStore', {
   state: (): {
-    StorePerformanceList?: StorePerformance[]
     todaySaleData?: todaySales // 今日本店销售数据
     TodayInventory?: TodayInventory // 今日库存数据
   } => ({
-    StorePerformanceList: undefined,
     todaySaleData: undefined,
     TodayInventory: undefined,
   }),
@@ -13,18 +11,10 @@ export const homeDataStore = defineStore('homeDataStore', {
 
   },
   actions: {
-    // 获取店铺列表销售数据
-    async StorePerformance(req: { duration: number }) {
-      this.StorePerformanceList = undefined
-      const { data } = await https.post<StorePerformance[], { duration: number }>('/statistic/store_sales_total', req)
-      if (data.value?.code === HttpCode.SUCCESS) {
-        this.StorePerformanceList = data.value.data
-      }
-    },
     // 获取今日本店销售数据
     async myStoreTodaySale(req: { store_id: string }) {
       this.todaySaleData = undefined
-      const { data } = await https.post<todaySales, { store_id: string }>('/statistic/today_sales', req)
+      const { data } = await https.post<todaySales, { store_id: string }>('/statistic/today/sales', req)
       if (data.value?.code === HttpCode.SUCCESS) {
         this.todaySaleData = data.value.data
       }
@@ -32,7 +22,7 @@ export const homeDataStore = defineStore('homeDataStore', {
     // 本店今日货品库存情况
     async myStoreTodayInventory(req: { store_id: string }) {
       this.TodayInventory = undefined
-      const { data } = await https.post<TodayInventory, { store_id: string }>('/statistic/today_product', req)
+      const { data } = await https.post<TodayInventory, { store_id: string }>('/statistic/today/product', req)
       if (data.value?.code === HttpCode.SUCCESS) {
         this.TodayInventory = data.value.data
       }
